feat(PhpAstExtractor): Add extraction of FormType implicit labels
